🔥: Remove service from pdf-manager that was not working
