+++ libs/api/pdf-manager/api/feature/src/lib/api-pdf-manager-api-feature.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ApiPdfManagerApiFeatureResolver } from './api-pdf-manager-api-feature.resolver';
 // import { ApiPdfManagerServiceFeatureService, ApiPdfManagerServiceFeatureModule } from '@conversation-catcher/api/pdf-manager/service/feature';
-import { ApiPdfManagerServiceModule } from '@conversation-catcher/api/pdf-manager/service';
+import { ApiPdfManagerServiceModule } from '@conversation-catcher/api/pdf-manager/service/feature';
 import { CqrsModule } from '@nestjs/cqrs';
 
 @Module({
