--- libs/api/pdf-manager/service/README.md
@@ -1,7 +1,0 @@
-# api-pdf-manager-service
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test api-pdf-manager-service` to execute the unit tests via [Jest](https://jestjs.io).
