+++ libs/api/pdf-manager/service/feature/jest.config.ts
@@ -1,6 +1,7 @@
+/* eslint-disable */
 export default {
-  displayName: 'api-pdf-manager-service-feature',
-  preset: '../../../../../jest.preset.js',
+  displayName: 'api-pdf-manager-service',
+  preset: '../../../../jest.preset.js',
   globals: {
     'ts-jest': {
       tsconfig: '<rootDir>/tsconfig.spec.json',
@@ -11,6 +12,5 @@ export default {
     '^.+\\.[tj]s$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory:
-    '../../../../../coverage/libs/api/pdf-manager/service/feature',
+  coverageDirectory: '../../../../coverage/libs/api/pdf-manager/service',
 };
