+++ libs/api/pdf-manager/service/feature/src/lib/api-pdf-manager-service.service.ts
