--- libs/api/pdf-manager/service/feature/src/lib/commands/events/delete-pdf-manager.event.handler.ts
@@ -1,9 +1,0 @@
-import { EventsHandler, IEventHandler } from "@nestjs/cqrs";
-import { DeletePdfEvent } from "./delete-pdf-manager.event";
-
-@EventsHandler(DeletePdfEvent)
-export class DeletePdfEventHandler  implements IEventHandler<DeletePdfEvent> {
-    handle(event: DeletePdfEvent) {
-        console.log('event has been handled: ', event);
-    }
-}
