--- libs/api/pdf-manager/service/feature/src/lib/commands/events/delete-pdf-manager.event.ts
@@ -1,7 +1,0 @@
-//import { FileCategory } from "@prisma/client";
-
-export class DeletePdfEvent  {
-    constructor(
-      public readonly id: string,
-    ) {}
-}
