--- libs/api/pdf-manager/service/feature/src/lib/commands/events/index.ts
@@ -1,4 +1,0 @@
-export * from './delete-pdf-manager.event';
-export * from './delete-pdf-manager.event.handler';
-export * from './set-pdf-manager.event';
-export * from './set-pdf-manager.event.handler';
