--- libs/api/pdf-manager/service/feature/src/lib/commands/events/set-pdf-manager.event.ts
@@ -1,14 +1,0 @@
-//import { FileCategory } from "@prisma/client";
-
-export class SetDownloadedPdfEvent  {
-    constructor(
-      public readonly id: string
-    ) {}
-}
-
-export class SetNamePdfEvent  {
-    constructor(
-      public readonly id: string,
-      public readonly name: string
-    ) {}
-}
