+++ libs/api/pdf-manager/service/feature/src/lib/commands/handlers/delete-pdf-manager.handler.ts
@@ -1,18 +1,13 @@
-import { HttpService } from '@nestjs/axios';
-import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
-import { PdfManagerServiceModel } from '../../models/pdf-manager-service-feature.model';
-//import {  } from "../events/delete-.event";
+import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { DeletePdfCommand } from '../impl/delete-pdf-manager.command';
 import { MongoDBAccess } from '@conversation-catcher/api/pdf-manager/repository/data-access';
 
 @CommandHandler(DeletePdfCommand)
 export class DeletePdfHandler implements ICommandHandler<DeletePdfCommand> {
-  constructor(
-    private publisher: EventPublisher,
-    private repository: MongoDBAccess
-  ) {}
+  constructor(private repository: MongoDBAccess) {}
 
   async execute({ id }: DeletePdfCommand) {
+    // console.log('Delete command called');
     return this.repository.deletePDF(id);
   }
 }
