+++ libs/api/pdf-manager/service/feature/src/lib/commands/handlers/index.ts
@@ -1,2 +1,2 @@
-export * from './delete-pdf-manager.handler';
-export * from './set-pdf-manager.handler';
+export * from './set-pdf-manager.handler'
+export * from './delete-pdf-manager.handler'
