+++ libs/api/pdf-manager/service/feature/src/lib/commands/impl/delete-pdf-manager.command.ts
@@ -1,7 +1,5 @@
-//import { FileCategory } from "@prisma/client";
-
 export class DeletePdfCommand {
-    constructor(
-      public readonly id?: string,
-    ) {}
+  constructor(
+    public readonly id?: string,
+  ) {}
 }
