+++ libs/api/pdf-manager/service/feature/src/lib/commands/impl/index.ts
@@ -1,2 +1,1 @@
-export * from './delete-pdf-manager.command';
-export * from './set-pdf-manager.command';
+export class Index {}
