+++ libs/api/pdf-manager/service/feature/src/lib/commands/impl/set-pdf-manager.command.ts
@@ -1,14 +1,20 @@
-//import { FileCategory } from "@prisma/client";
-
 export class SetDownloadedPdfCommand {
-    constructor(
-      public readonly id?: string,
-    ) {}
+  constructor(
+    public readonly id?: string,
+  ) {}
 }
 
 export class SetNamePdfCommand {
-    constructor(
-      public readonly id?: string,
-      public readonly name?: string,
-    ) {}
+  constructor(
+    public readonly id?: string,
+    public readonly name?: string,
+  ) {}
+}
+
+export class AddPdfCommand {
+  constructor(
+    public readonly email?: string,
+    public readonly name?: string,
+    public readonly text?: string,
+  ) {}
 }
