--- libs/api/pdf-manager/service/feature/src/lib/models/index.ts
@@ -1,1 +1,0 @@
-export * from './pdf-manager-service-feature.model';
