--- libs/api/pdf-manager/service/feature/src/lib/models/pdf-manager-service-feature.model.ts
@@ -1,21 +1,0 @@
-import { AggregateRoot } from "@nestjs/cqrs";
-//import { FileCategory } from "@prisma/client";
-
-/*
-id
-name
-path
-creationDate
-downloaded
-*/
-
-export class PdfManagerServiceModel extends AggregateRoot {
-    constructor() {
-      super();
-    }
-    id?: string;
-    name?: string;
-    path?: string;
-    creationDate?: Date;
-    downloaded?: boolean;
-  }
