+++ libs/api/pdf-manager/service/feature/src/lib/queries/handlers/get-pdf-manager.handler.ts
