+++ libs/api/pdf-manager/service/feature/src/lib/queries/handlers/index.ts
@@ -1,1 +1,1 @@
-export * from './get-pdf-manager.handlers';
+export * from './get-pdf-manager.handler'
