+++ libs/api/pdf-manager/service/feature/src/lib/queries/impl/index.ts
@@ -1,1 +1,1 @@
-export * from './get-pdf-manager.query';
+export * from './get-pdf-manager.query'
