--- libs/api/pdf-manager/service/jest.config.ts
@@ -1,16 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'api-pdf-manager-service',
-  preset: '../../../../jest.preset.js',
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json',
-    },
-  },
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.[tj]s$': 'ts-jest',
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../../../coverage/libs/api/pdf-manager/service',
-};
