--- libs/api/pdf-manager/service/project.json
@@ -1,23 +1,0 @@
-{
-  "$schema": "..\\..\\..\\..\\node_modules\\nx\\schemas\\project-schema.json",
-  "sourceRoot": "libs/api/pdf-manager/service/src",
-  "projectType": "library",
-  "targets": {
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/api/pdf-manager/service/**/*.ts"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/api/pdf-manager/service"],
-      "options": {
-        "jestConfig": "libs/api/pdf-manager/service/jest.config.ts",
-        "passWithNoTests": true
-      }
-    }
-  },
-  "tags": []
-}
