--- libs/api/pdf-manager/service/src/index.ts
@@ -1,2 +1,0 @@
-export * from './lib/api-pdf-manager-service.service';
-export * from './lib/api-pdf-manager-service.module';
