--- libs/api/pdf-manager/service/src/lib/commands/handlers/index.ts
@@ -1,2 +1,0 @@
-export * from './set-pdf-manager.handler'
-export * from './delete-pdf-manager.handler'
