--- libs/api/pdf-manager/service/src/lib/commands/impl/delete-pdf-manager.command.ts
@@ -1,5 +1,0 @@
-export class DeletePdfCommand {
-  constructor(
-    public readonly id?: string,
-  ) {}
-}
