--- libs/api/pdf-manager/service/src/lib/commands/impl/index.ts
@@ -1,1 +1,0 @@
-export class Index {}
