--- libs/api/pdf-manager/service/src/lib/queries/impl/get-pdf-manager.query.ts
@@ -1,7 +1,0 @@
-export class GetPdfByIdQuery {
-  constructor(public readonly id) {}
-}
-
-export class GetPdfsQuery {
-  constructor(public readonly userid) {}
-}
