--- libs/api/pdf-manager/service/src/lib/queries/impl/index.ts
@@ -1,1 +1,0 @@
-export * from './get-pdf-manager.query'
