+++ libs/api/shell/src/lib/api-shell.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common';
 import { GraphQLModule } from '@nestjs/graphql';
 import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo';
-import { ApiDatabaseManagerApiFeatureModule as DatabaseManager } from '@conversation-catcher/api/database-manager/api/feature';
 import { ApiStudentExploreApiFeatureModule as PdfManager } from '@conversation-catcher/api/pdf-manager/api/feature';
 import { ApiSummariseTextApiModule as Summariser} from '@conversation-catcher/api/summarise-text/api'
 import { ApiSpeechToTextApiModule as SpeechToText } from '@conversation-catcher/api/speech-to-text/api';
@@ -16,7 +15,6 @@ import { ApiSpeechToTextApiModule as SpeechToText } from '@conversation-catcher/
       driver: ApolloDriver,
       playground: true,
     }),
-    DatabaseManager,
     PdfManager,
     Summariser,
     SpeechToText
