+++ workspace.json
@@ -8,7 +8,6 @@
     "api-pdf-manager-api-data-access": "libs/api/pdf-manager/api/data-access",
     "api-pdf-manager-api-feature": "libs/api/pdf-manager/api/feature",
     "api-pdf-manager-repository-data-access": "libs/api/pdf-manager/repository/data-access",
-    "api-pdf-manager-service": "libs/api/pdf-manager/service",
     "api-pdf-manager-service-feature": "libs/api/pdf-manager/service/feature",
     "api-pdf-manager-shared": "libs/api/pdf-manager/shared",
     "api-shell": "libs/api/shell",
